fix compilation (ull -> uint64_t)

+++ include/gpu_util.cuh
@@ -15,8 +15,8 @@ void cudaCallBlockCount(
         const uint words_per_vector,
         const uint vectors_per_config,
         const uint configs_per_query,
-        const unsigned long long** bvectorsPtr,
-        unsigned long long* results,
-        unsigned long long* intermediateStatesPtr);
+        const uint64_t** bvectorsPtr,
+        uint64_t* results,
+        uint64_t* intermediateStatesPtr);
 
 #endif
